requirements.txt
pandas
numpy
python-dotenv
pymongo
certifi
pymongo[srv]
pyaml
dill
scikit-learn
-e .